+++ R/loo.R
@@ -481,7 +481,7 @@ loo_compare <- function(x, ...) UseMethod("loo_compare")
 #' comparison table. Sampled versus marginalized local-effect provenance does
 #' not by itself block comparison.
 #'
-#' @return A matrix of class \code{"compare.loo"} as returned by
+#' @return A data frame of class \code{"compare.loo"} as returned by
 #' \code{\link[loo]{loo_compare}}.
 #'
 #' @seealso \code{\link{loo.brma}}, \code{\link[loo]{loo_compare}}
@@ -544,7 +544,7 @@ loo_compare.brma <- function(x, ..., unit = "estimate") {
 #' objects to compare.
 #' @param unit output/deletion unit used when extracting LOO from brma objects.
 #'
-#' @return A matrix of class \code{"compare.loo"} as returned by
+#' @return A data frame of class \code{"compare.loo"} as returned by
 #' \code{\link[loo]{loo_compare}}.
 #'
 #' @seealso \code{\link{loo.brma}}, \code{\link[loo]{loo_compare}}